% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_new_slide_utils.R,
%   R/slide_modifications.R, R/update_slide.R
\name{add_slide}
\alias{add_slide}
\alias{delete_slide}
\alias{move_slide}
\alias{update_slide}
\title{Working with Slides}
\usage{
add_slide(
  pptx,
  layout,
  ...,
  index,
  transition = NULL,
  polish_error_continue = TRUE,
  error_call = current_env()
)

delete_slide(pptx, index, error_call = current_env())

move_slide(pptx, index, to, error_call = current_env())

update_slide(
  pptx,
  index,
  ...,
  transition = NULL,
  polish_error_continue = TRUE,
  error_call = current_env(),
  verbose = getOption("mirage.verbose", default = TRUE)
)
}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{layout}{slide layout to use when creating a slide in \code{add_slide()}}

\item{...}{content objects created with \code{\link[=content]{content()}}}

\item{index}{slide index. By default \code{add_slide()} creates a new slide at the
end of the slide deck.}

\item{transition}{Defines the type of transition to use when moving to this slide. See \code{?slide_transitions} to see viable values.}

\item{polish_error_continue}{Boolean indicating if when an error is throwing
in either polishing or converting the placeholder if it should be
propagated as an error (FALSE) or just inform the user and continue (TRUE).
Defaults to TRUE.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{to}{new index to move the slide to}

\item{verbose}{if TRUE, some information is \code{\link[cli:cli_inform]{cli::cli_inform()}}ed along the
way}
}
\value{
a \code{pptx_container} object with the slide added, updated, moved, or deleted
}
\description{
\code{add_slide()} Adds a new slide at a specified position, and
populate it with content from \code{...}.

\code{update_slide()} updates an existing slide from a specific index and populate
it with content from \code{...}.

\code{move_slide()} moves the slide from one index to another.

\code{delete_slide()} deletes the slide based on the index.
}
\examples{

library(mirage)

## Create a new presentation
ppt <- load_pptx()

## adding new slides
ppt <- ppt |>
  ## add first slide with content in title
  add_slide(layout = "Title and Content", 
    content("hello world", ph = ph_title())
  ) |>
  ## add another slide with nothing on it yet
  add_slide(layout = "Title and Content") |> 
  ## slide added at index 2 with content in title and body
  add_slide(layout = "Title and Content", 
    index = 2, 
    content("Hello World", ph = ph_title()), 
    content("This is a body", ph = ph_body())
  )
 
## updating an existing slide wit new content
ppt <- ppt |>
   update_slide(index = 1, 
     content("This placeholder now has a value", ph = ph_body())
   )

## moving a slide from index 2 to index 1
ppt <- ppt |>
   move_slide(index = 2, to = 1)

## deleting a slide at index 3
ppt <- ppt |>
   delete_slide(index = 3)

}
