% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{args_mirage}
\alias{args_mirage}
\title{Documentation anchor for mirage functions}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}
}
\description{
Use \verb{@inheritParams mirage::args_mirage} to document \code{pptx} for your functions
}
