% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_version.R
\name{check_pkg_version}
\alias{check_pkg_version}
\title{check package versions}
\usage{
check_pkg_version(pkg, version)
}
\arguments{
\item{pkg}{package}

\item{version}{version that is expected}
}
\description{
check package versions and inform if they don't match
the version specified
}
