% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_metadata_element}
\alias{extract_metadata_element}
\title{Extract metadata element}
\usage{
extract_metadata_element(metadata, name = "name", empty = "<display>")
}
\arguments{
\item{metadata}{named character vector}

\item{name}{name of element to extract}

\item{empty}{synonyms for ""}
}
\value{
the value of the metadata element, or "" if not found
}
\description{
Extract metadata element
}
\examples{
m <- c(foo = "abc", bar = "def")
extract_metadata_element(m, name = "foo")
extract_metadata_element(m, name = "bar", empty = "def") ## return ""

}
