% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{get_layout_names}
\alias{get_layout_names}
\title{Get names for all the the slides}
\usage{
get_layout_names(pptx, error_call = current_env())
}
\arguments{
\item{pptx}{Powerpoint file}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
character vector of layout names
}
\description{
Get names for all the the slides
}
