% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{get_slide_file_name}
\alias{get_slide_file_name}
\title{Get the file name for a given slide}
\usage{
get_slide_file_name(pptx, index, error_call = current_env())
}
\arguments{
\item{pptx}{A PowerPoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{index}{slide index}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
The file name for the slide, as a character string
}
\description{
Get the file name for a given slide
}
\examples{
pptx <- example_pptx()
get_slide_file_name(pptx, index = 1)

}
