% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_layouts.R
\name{list_layouts}
\alias{list_layouts}
\title{List layouts}
\usage{
list_layouts(pptx, error_call = current_env())
}
\arguments{
\item{pptx}{A PowerPoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A tibble with information about the layouts
}
\description{
List layouts
}
\examples{
list_layouts(load_pptx())

}
