% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{extended properties metadata}
\usage{
metadata(data, uri = "r://package/powerup", error_call = current_env())
}
\arguments{
\item{data}{Named character vector used as name= and value= for the \code{custom:property} nodes.}

\item{uri}{uri, used in the \code{p:ext} node}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
a \code{p:extLst} tag
}
\description{
extended properties metadata
}
\examples{
metadata(c(foo = "abc", bar = "def"))

}
