% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder_templates.R
\name{templated_placeholder}
\alias{templated_placeholder}
\alias{ph_title}
\alias{ph_subtitle}
\alias{ph_body}
\alias{ph_body_left}
\alias{ph_body_right}
\alias{ph_footer}
\alias{ph_footer_left}
\alias{ph_footer_right}
\title{Placeholder Templates}
\usage{
templated_placeholder(preferred, fallback, error_call = current_env())

ph_title(replace = TRUE)

ph_subtitle(replace = TRUE)

ph_body(replace = TRUE)

ph_body_left(replace = TRUE)

ph_body_right(replace = TRUE)

ph_footer(replace = TRUE)

ph_footer_left(replace = TRUE)

ph_footer_right(replace = TRUE)
}
\arguments{
\item{preferred}{placeholder object of class \verb{existing placeholder}}

\item{fallback}{placeholder object of class \code{new_placeholder}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{replace}{Identify whether to replace the existing placeholder if
it exists on the slide. Defaults to TRUE.}
}
\value{
a \code{templated_placeholder} object
}
\description{
Included standard template locations to put standard contents when
constructing a slide.

These are all "templated_placeholder" objects, which allows the user to
specify two potential placeholders, allowing passing a \code{existing_placeholder}
object( \code{ref_placeholder()}), which tries to use an existing placeholder
either in the slide or defined in the layout, and the fallback must be a
\code{new_placeholder} object (\code{new_placeholder()}), which defines a placeholder
into the slide without needing there to be a pre-defined placeholder in the
slide layout.
}
\examples{
## create a new placeholder with a label of "my_placeholder"
templated_placeholder(
  preferred = ref_placeholder(label = "title", type = "title"),
  fallback = new_placeholder(label = "my_placeholder", type = "body", x_offset = "5\%", width = "90\%", y_offset = "5\%", height = "12.5\%")
)
 
}
