% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_placeholder.R
\name{process_placeholder}
\alias{process_placeholder}
\title{Internal method for processing placeholder requests}
\usage{
process_placeholder(ph, pptx, index, ..., error_call = current_env())
}
\arguments{
\item{ph}{placeholder}

\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{index}{Slide number}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Internal method for processing placeholder requests
}
