% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{read_slide_layout}
\alias{read_slide_layout}
\title{Read the layout for a given slide}
\usage{
read_slide_layout(pptx, index, error_call = current_env())
}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{index}{slide index}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
The slide layout, as read by \link[xml2:read_xml]{xml2::read_xml}
}
\description{
Read the layout for a given slide
}
