% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{read_slide_placeholders}
\alias{read_slide_placeholders}
\title{Read placeholders information from PowerPoint presentation}
\usage{
read_slide_placeholders(
  pptx,
  index,
  only_slide = FALSE,
  only_layout = FALSE,
  keep_all = FALSE,
  error_call = current_env()
)
}
\arguments{
\item{pptx}{A PowerPoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{index}{slide index}

\item{only_slide}{if TRUE only extract placeholders from the slide, i.e. not from its associated layout}

\item{only_layout}{if TRUE only extract placeholders from the layout, i.e. not occupied in the slide}

\item{keep_all}{if FALSE, only keep non missing placeholders}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Read placeholders information from PowerPoint presentation
}
\examples{
ppt <- example_pptx()
read_slide_placeholders(ppt)
read_slide_placeholders(ppt, index = 1)
read_slide_placeholders(ppt, index = 1:2)

read_slide_placeholders(ppt, only_slide = TRUE)
read_slide_placeholders(ppt, index = 2, only_slide = TRUE)

}
