% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_transitions.R
\name{slide_transition_generic}
\alias{slide_transition_generic}
\alias{slide_transition_blinds}
\alias{slide_transition_checker}
\alias{slide_transition_circle}
\alias{slide_transition_comb}
\alias{slide_transition_cover}
\alias{slide_transition_cut}
\alias{slide_transition_diamond}
\alias{slide_transition_dissolve}
\alias{slide_transition_fade}
\alias{slide_transition_newsflash}
\alias{slide_transition_plus}
\alias{slide_transition_pull}
\alias{slide_transition_push}
\alias{slide_transition_random}
\alias{slide_transition_randomBar}
\alias{slide_transition_split}
\alias{slide_transition_wedge}
\alias{slide_transition_wheel}
\alias{slide_transition_wipe}
\alias{slide_transition_zoom}
\alias{slide_transition_null}
\title{Defined Slide Transitions}
\usage{
slide_transition_generic(
  type,
  transition_type_xml,
  speed = c("med", "fast", "slow")
)

slide_transition_blinds(orientation = c("horz", "vert"), speed = "med")

slide_transition_checker(orientation = c("horz", "vert"), speed = "med")

slide_transition_circle(orientation = c("horz", "vert"), speed = "med")

slide_transition_comb(orientation = c("horz", "vert"), speed = "med")

slide_transition_cover(
  direction = c("d", "l", "r", "u", "ld", "lu", "rd", "ru"),
  speed = "med"
)

slide_transition_cut(black_screen = FALSE, speed = "med")

slide_transition_diamond(speed = "med")

slide_transition_dissolve(speed = "med")

slide_transition_fade(black_screen = FALSE, speed = "med")

slide_transition_newsflash(speed = "med")

slide_transition_plus(speed = "med")

slide_transition_pull(direction = c("ld", "lu", "rd", "ru"), speed = "med")

slide_transition_push(direction = c("d", "l", "r", "u"), speed = "med")

slide_transition_random(speed = "med")

slide_transition_randomBar(orientation = c("horz", "vert"), speed = "med")

slide_transition_split(
  split = c("in", "out"),
  orientation = c("horz", "vert"),
  speed = "med"
)

slide_transition_pull(direction = c("ld", "lu", "rd", "ru"), speed = "med")

slide_transition_wedge(speed = "med")

slide_transition_wheel(spokes = 4L, speed = "med")

slide_transition_wipe(direction = c("d", "l", "r", "u"), speed = "med")

slide_transition_zoom(zoom = c("in", "out"), speed = "med")

slide_transition_null()
}
\arguments{
\item{type}{string defining the slide transition type}

\item{transition_type_xml}{string defining the 'OOXML' of the transition type}

\item{speed}{speed of the slide transition - med, fast, or slow}

\item{orientation}{specify "horz" or "vert", defining the orientation of a
transition}

\item{direction}{specify the direction of transition, with possible values of
d (down), l (left), r (right), u (up), ld (left down), lu (left up), rd
(right down), ru (right up). Not every transition allows every direction.
Check the available default values to see what values can be used for every
transition}

\item{black_screen}{boolean identifying if  the transition starts from a
black screen and then transitions to new slide over black}

\item{spokes}{specify the number of spokes in the wheel; values are integers.}

\item{zoom, split}{specify the direction of zoom, "in" or "out"}
}
\value{
a \code{slide_transition} object for the defined transition, or a \code{slide_transition_null} object for no transition or removing a transition. The \code{slide_transition} object can be added to a slide via \code{\link[=add_slide]{add_slide()}} or \code{\link[=update_slide]{update_slide()}}.
}
\description{
Slide transitions define the transition to show when moving to the slide.
\code{\link[=slide_transition_generic]{slide_transition_generic()}} allows the user to create and pass new types of
transitions. Otherwise, most standard transitions are already defined as
\verb{slide_transition_*()} to allow the user to easily define a transtion and any
other assets the transition may need.
}
\examples{

pptx <- example_pptx()

checker_transition <- slide_transition_checker()
dissolve_transition <- slide_transition_dissolve()

## add a transition to a new slide
pptx <- pptx |>
   add_slide(
     layout = "Title and Content",
     transition = checker_transition
   )

## Update/add transition to n existing slide
pptx <- pptx |>
   update_slide(
     index = 1,
     transition = dissolve_transition
   )

}
