% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_style.R
\name{add_table_style}
\alias{add_table_style}
\title{Add Table Style}
\usage{
add_table_style(
  pptx,
  table_style,
  verbose = getOption("mirage.verbose", default = FALSE),
  error_call = caller_env()
)
}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{table_style}{an xml_node that defines a tblStyle. Created by \code{\link[=new_table_style]{new_table_style()}}.}

\item{verbose}{if TRUE, some information is \code{\link[cli:cli_abort]{cli::cli_inform()}}ed along the way}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Given a provided table style, add it to the tableStyle file in the document.
this makes it available to tables that get added to the powerpoint as a table
style option
}
