% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{args_powerup}
\alias{args_powerup}
\title{Documentation anchor for powerup functions}
\usage{
N/A - This documentation anchor is not intended to be called directly.
}
\arguments{
\item{pptx}{A PowerPoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}
}
\description{
Use \verb{@inheritParams powerup::args_powerup} to document \code{pptx} for your functions
}
\keyword{internal}
