% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content}
\alias{content}
\alias{remove_content}
\title{Add or remove content from a slide}
\usage{
content(
  value,
  ph = ph_body(),
  metadata = NULL,
  ...,
  group_contents = TRUE,
  error_call = current_env()
)

remove_content(ph, error_call = current_env())
}
\arguments{
\item{value}{value to be polished and then added at the placeholder defined}

\item{ph}{Defining where values should be added to or removed from. Must be a
placeholder object}

\item{metadata}{named character vector of metadata to store in the
\code{p:extLst/p:ext/custom:meta/custom:metadata} nodes}

\item{...}{arguments to include when polishing \code{value}}

\item{group_contents}{boolean indicating if the polished value should be
shown in the powerpoint as a grouped Sp. The default, and for most cases,
TRUE.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=content]{content()}} lets a user specify a value to add to a slide, and the location
on the slide to add it through the ph argument. ph defines the placeholder,
or height, width, x and y offset of the bounding box where the content should
be added.
}
\details{
\code{\link[=remove_content]{remove_content()}} on the other hand, lets users specify a placeholder with
content that should be removed from the slide using the same placeholder
nomenclature. The placeholder must exist in the slide to be removed.
}
\examples{

content("My Title 2", ph = ph_title())
content(mtcars      , ph = ph_body())

remove_content(ph = ph_title())
remove_content(ph = ref_placeholder(label = "An Existing Placeholder"))

}
