% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_metadata_element}
\alias{extract_metadata_element}
\title{Extract metadata element}
\usage{
extract_metadata_element(metadata, name = "name", empty = "<display>")
}
\arguments{
\item{metadata}{named character vector}

\item{name}{name of element to extract}

\item{empty}{synonyms for ""}
}
\description{
Extract metadata element
}
