% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_style.R
\name{get_table_styles}
\alias{get_table_styles}
\title{List Table Styles}
\usage{
get_table_styles(pptx)
}
\arguments{
\item{pptx}{A PowerPoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}
}
\value{
A data frame with columns \code{style_id} and \code{style_name} for each table style in the presentation.
}
\description{
List the available table styles within the document
}
\examples{
pptx <- example_pptx()
get_table_styles(pptx)

}
