% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_style.R
\name{get_table_styles}
\alias{get_table_styles}
\title{List Table Styles}
\usage{
get_table_styles(pptx)
}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}
}
\description{
List the available table styles within the document
}
