% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{list_placeholders}
\alias{list_placeholders}
\title{List the available placeholders in a slide or a given layout}
\usage{
list_placeholders(
  pptx,
  index = NULL,
  layout = NULL,
  units = c("in", "cm"),
  keep_all = FALSE,
  error_call = current_env()
)
}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{index}{Slide index}

\item{layout}{the name of the layout in the powerpoint. Overridden if index is provided}

\item{units}{"in" or "cm", what units should be used for the values presented
(offsets, dimensions)}

\item{keep_all}{Keep all the placeholders (TRUE), or only the ons with a \if{html}{\out{<ph>}} node (FALSE)}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
List the available placeholders in a slide or a given layout
}
\examples{

pptx <- example_pptx()

# what are the available placeholders on the slides
list_placeholders(pptx, layout = "Two Content")
list_placeholders(pptx, index = 2)

}
