% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_slides.R
\name{list_slides}
\alias{list_slides}
\title{List slides}
\usage{
list_slides(pptx, keep_all = FALSE, error_call = current_env())
}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{keep_all}{if FALSE, only keep non missing placeholders}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A tibble with information about the slides
}
\description{
List slides
}
\examples{
list_slides(load_pptx())

}
