% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirage.R
\name{mirage}
\alias{mirage}
\title{Mixture model based rare variant analysis}
\usage{
mirage(...)
}
\arguments{
\item{data}{geno-pheno data  which is a list, of which every element is also a list with the first element is 
a genotype matrix with every row being an individual and each column of a variant, the second element is phenotype.}

\item{partition}{Variants partitions. For every gene, which variants belong to first variant category and which 
variants belong to the second, and so on.}

\item{gamma}{a vector of category specific effect size prior}

\item{sigma}{prior on proportion of risk genes}
}
\value{
\item{BF}{Bayes factor of genes}
\item{delta}{...}
\item{eta}{...}
\item{pvalue}{...}
}
\description{
This function implements the gene based version of rare variant test
with MIRAGE model
}
\examples{
mirage(...)
}
