% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirage.R
\name{mirage}
\alias{mirage}
\title{mirage: MIxture model based Rare variant Analysis on GEnes}
\usage{
mirage(data, n1, n2, gamma = 3, sigma = 2, eta.init = 0.1,
  delta.init = 0.1, estimate.delta = TRUE, max.iter = 10000,
  tol = 1e-05, verbose = TRUE)
}
\arguments{
\item{data}{variant count data, a 5 column data frame for 1) locus ID, 2) Gene  3) count in cases, 
4) count in controls and 5) variant category index for a variant. The 1st column is optional.}

\item{n1}{sample size in cases.}

\item{n2}{sample size in controls.}

\item{gamma}{a list of category specific hyper prior shape parameter in Beta distribution  for effect size, or a numeric value if all category share the same effect size.}

\item{sigma}{a list of category specific hyper prior scale parameter in Beta distribution for effect size, or a numeric value if all category share the same effect size.}

\item{eta.init}{initial value for prior on proportion of risk variants in a variant set.}

\item{delta.init}{initial value for prior on proportion of risk genes. Must be a positive number between 0 and 1.}

\item{estimate.delta}{When TRUE delta is to be estimated and FALSE delta is fixed at delta.init}

\item{max.iter}{maximum number of iterations enforcing EM algorithm to stop}

\item{tol}{threshold of parameter estimate difference to determine the convergence of EM algorithm}
}
\value{
\item{BF.gene}{Bayes factor of genes}
\item{delta.est}{Estimate for proportion of risk genes}
\item{delta.pvalue}{Significant test for delta = 0}
\item{eta.est}{Estimate for proportion of risk variants in a variant group}
\item{eta.pvalue}{Significant test for eta = 0}
\item{BF.all}{a list of Bayes factor for all variants in a gene}
}
\description{
This function implements rare variant test with full MIRAGE model with both variant and gene level
}
\examples{
# see example at https://xinhe-lab.github.io/mirage/articles/mwe.html
}
