% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_style.R
\name{new_table_style}
\alias{new_table_style}
\title{Create a new table style node}
\usage{
new_table_style(
  style_name,
  style_id = make_style_id(),
  text_color = "#000000",
  cell_color = "#f36633",
  border_color = "#959595",
  border_style = "sng",
  border_weight = 12700,
  row_header_text_color = text_color,
  row_total_text_color = text_color,
  col_first_text_color = text_color,
  col_last_text_color = text_color,
  row_header_cell_color = cell_color,
  row_total_cell_color = cell_color,
  col_first_cell_color = cell_color,
  col_last_cell_color = cell_color,
  cell_color_tint1 = tint_color(cell_color, tint = 0.4),
  cell_color_tint2 = tint_color(cell_color, tint = 0.2),
  major_border_color = border_color,
  major_border_style = border_style,
  major_border_weight = border_weight * 3,
  error_call = caller_env()
)
}
\arguments{
\item{style_name}{The new name of the style. This is what will be referenced
when defining the table style..}

\item{style_id}{a unique string containing 8, 4, 4, 4, then 12 random values
between A-F and 0-9, separated by dashes:
\verb{\\\\\{[0-9A-F]\{8\}-[0-9A-F]\{4\}-[0-9A-F]\{4\}-[0-9A-F]\{4\}-[0-9A-F]\{12\}\\\\\}}}

\item{text_color, cell_color, border_color, row_header_text_color, row_total_text_color, col_first_text_color, col_last_text_color, row_header_cell_color, row_total_cell_color, col_first_cell_color, col_last_cell_color, cell_color_tint1, cell_color_tint2, major_border_color}{either a color name (as listed by colors()), a hexadecimal string}

\item{border_style, major_border_style}{Define how borders should be styled.
Options include: "sng","dbl","thickThin","thinThick","tri"}

\item{border_weight, major_border_weight}{Define the thickness of the borders.
Must be a value between 0 and 20116800.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An xml_node that defines a tblStyle. Added to a powerpoint via add_table_style().
}
\description{
Helper function to make it simple to define a table style node.
}
\examples{
example_table_style <- new_table_style("Example Style", text_color = "blue", border_color = "pink")

}
