% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{ref_placeholder}
\alias{ref_placeholder}
\alias{new_placeholder}
\title{Placeholder object}
\usage{
ref_placeholder(
  label = "",
  label_match = c("exact", "match"),
  label_from = c("slide", "layout"),
  type = c("body", "title", "ctrTitle", "ftr", "subTitle", "tbl", "chart", "img", "any"),
  tie_breaker = c("largest", "smallest", "left", "right", "top", "bottom"),
  replace = TRUE,
  error_call = current_env()
)

new_placeholder(
  type = c("body", "title", "ctrTitle", "ftr", "subTitle", "tbl", "chart", "img"),
  x_offset = "10\%",
  y_offset = "20\%",
  width = "80\%",
  height = "60\%",
  label = "",
  error_call = current_env()
)
}
\arguments{
\item{label}{The placeholder label to select. \code{ref_placeholder} allows to
pass a pattern to search with if label_match is set to "match".}

\item{label_match}{Use "match" if \code{label} is a regex pattern, or "exact"
for exact match. Defaults to "exact".}

\item{label_from}{Which placeholder label should be used, the one listed in
the "slide" or from the "layout". Generally new slides should use "layout",
while updating a slide may use either. Listing both allows to
check the first listed location then the other before declaring it cannot
find it.}

\item{type}{The type of placeholder. Options are "body", "title", "ftr",
"subTitle", "tbl", "chart", "img".
\code{ref_placeholder()} allows for "any" in
this position to allow for referencing any type of placeholder, or any
combination of to be types to be passed. \code{new_placeholder()} allows only
one.}

\item{tie_breaker}{Which pane to select when multiple are returned from label
matching ("largest","left","right","top","bottom"). can pass multiple to
clarify if there are ties remaining.}

\item{replace}{Identify whether to replace the existing placeholder or not if
it exists on the slide. defaults to TRUE}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{x_offset}{\code{x} offset in centimeters (cm), inches (in), or percent(in) of slide width}

\item{y_offset}{\code{y} offset in centimeters, inches, or percent of slide height}

\item{width}{Width in centimeters, inches, or percent of slide width}

\item{height}{Height in centimeters, inches, or percent of slide height}
}
\value{
A placeholder object
}
\description{
Object to define placeholders.
}
\details{
\code{ref_placeholder()} defines what logic to apply when figuring out which a
placeholder in a slide or layout to use to insert content.

\code{new_placeholder()} defines what logic to use when creating a
new placeholder in a slide.

\code{empty_placeholder()} indicates that the content
should not be added to the slide, and does not expect any content.
}
\examples{

# reference placeholder, find the top-most placeholder in the slide or layout
# with the word "title" in the placeholder label where type is "title" and
# replace the contents when new values are inserted
ref_placeholder(
    label = "[Tt]itle",
    type = "title",
    tie_breaker = "top",
    label_match = "match",
    label_from = c("slide","layout"),
    replace = TRUE
    )

# reference placeholder, find the the exact placeholder with the label "Content Placeholder 5"
# based on the template layout
ref_placeholder(
    label = "Content Placeholder 5",
    label_match = "exact",
    label_from = "layout"
    )

# Create a new placeholder 20\% down from the top of the slide, 10\% off from the left side, and
# make it 10cm tall and 15cm wide
new_placeholder(
    x_offset = "10\%",
    y_offset = "20\%",
    height = "10cm",
    width = "15cm"
    )

}
