% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{polish_content_pptx.preset_shape}
\alias{polish_content_pptx.preset_shape}
\title{Polish fit models}
\usage{
\method{polish_content_pptx}{preset_shape}(x, ph = "<p:ph/>", pptx, ..., error_call = current_env())
}
\arguments{
\item{x}{preset_shape object to polish}

\item{ph}{placeholder}

\item{pptx}{The powerpoint presentation.}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Polish fit models
}
\examples{

pptx <- example_pptx() |>
 add_slide(
  layout = "Title and Content",
  content(shape("rightArrow", fill = "#000000", outline = "#ffffff"), ph = ph_title()),
  content(shape("heart", fill = "#ffffff", outline = "#000000"), ph = ph_body()),
  content(
    shape("funnel"),
    ph = new_placeholder(x_offset = "10\%",y_offset="20\%", height = "20\%",width="22\%")
   )
 )

file <- tempfile(fileext = ".pptx")
pptx |> save_pptx(file)

}
