% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_container.R
\name{load_pptx}
\alias{load_pptx}
\alias{slide_errors}
\alias{save_pptx}
\alias{example_pptx}
\title{Load/Save a PowerPoint file}
\usage{
load_pptx(path = NULL, name)

slide_errors(pptx)

save_pptx(pptx, path, error_call = current_env())

example_pptx()
}
\arguments{
\item{path}{output path of file to be saved}

\item{name}{name of the file}

\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
\code{load_pptx()} returns a new \code{pptx_container} object that wraps objects
created by \code{\link[officer:read_pptx]{officer::read_pptx()}} with custom functionality for mirage.
}
\description{
Load/Save a PowerPoint file
}
\details{
\code{slide_errors()} Extract a tibble of errors that occured when adding and updateing slides
}
\examples{

pptx <- load_pptx(system.file("ref_files/standard_powerpoint.pptx", package = "mirage"))

pptx2<- load_pptx()

}
