% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{read_layout_placeholders}
\alias{read_layout_placeholders}
\title{Read layout placeholders}
\usage{
read_layout_placeholders(
  pptx,
  layout,
  keep_all = FALSE,
  error_call = current_env()
)
}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{layout}{layout}

\item{keep_all}{if FALSE, only keep non missing placeholders}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Read layout placeholders
}
