% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{remove_slide_ph}
\alias{remove_slide_ph}
\title{Remove placeholder from a slide}
\usage{
remove_slide_ph(
  pptx,
  ph_id,
  index,
  error_call = current_env(),
  verbose = getOption("mirage.verbose", default = FALSE)
)
}
\arguments{
\item{pptx}{A Powerpoint wrapped in a <pptx_container> object. see \code{\link[=load_pptx]{load_pptx()}}}

\item{ph_id}{placeholder id}

\item{index}{index of slide}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{verbose}{inform about the removed placeholders}
}
\description{
Remove placeholder from a slide
}
