% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_style.R
\name{tint_color}
\alias{tint_color}
\title{Apply tint to a provided color}
\usage{
tint_color(x, tint)
}
\arguments{
\item{x}{either a color name (as listed by colors()), a hexadecimal string}

\item{tint}{the percentage to tint the color. Must be a value between 0 and 1}
}
\description{
Apply tint to a provided color
}
